reviews: Use default name
